% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_branch_vals.R
\name{assign_branch_vals}
\alias{assign_branch_vals}
\alias{assign_branch_vals.costeffectiveness_tree}
\title{Assign Branching Values to Decision Tree}
\usage{
assign_branch_vals(osNode, parameter_p, parameter_val)

\method{assign_branch_vals}{costeffectiveness_tree}(osNode, parameter_p = NA,
  parameter_val = NA)
}
\arguments{
\item{osNode}{data.tree object}

\item{parameter_p}{Decision tree node probabilities}

\item{parameter_val}{Decision tree node value}
}
\description{
Used in deterministic sensitivity analysis.
}
\examples{

}
