% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_branch_values.R
\name{assign_branch_values}
\alias{assign_branch_values}
\title{Wrapper to Assign Branching Values to Decision Tree}
\usage{
assign_branch_values(osNode.cost, osNode.health, parameter_p = NA,
  parameter_cost = NA)
}
\arguments{
\item{osNode.cost}{costeffectiveness.tree object}

\item{osNode.health}{costeffectiveness.tree object}

\item{parameter_p}{Decision tree node probabilities}

\item{parameter_cost}{Decision tree node values}
}
\description{
THIS REPLACES NON-S3 FUNCTION
}
\details{
Used in deterministic sensitivity analysis.
}
