% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expectedValues.R
\name{calc.expectedValues}
\alias{calc.expectedValues}
\title{Calculate Expected Values for Each Node of Decision Tree}
\usage{
calc.expectedValues(osNode)
}
\arguments{
\item{osNode}{}
}
\value{
osNode
}
\description{
Takes an object of class costeffectiveness.tree.
}
\examples{
## read-in decision tree
osNode <- create.costeffectiveness.tree(yaml_tree = "raw data/LTBI_dtree-cost-distns.yaml")
print(osNode, "type", "p", "distn", "mean", "sd")

## calculate a single realisation expected values
osNode <- calc.expectedValues(osNode)
print(osNode, "type", "p", "distn", "mean", "sd", "payoff")

## calculate multiple realisation for specific nodes
MonteCarlo.expectedValues(osNode, n=100)

}
\seealso{
create.costeffectiveness.tree, payoff
}

