% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pathway_probs.R
\name{calc.pathway_probs}
\alias{calc.pathway_probs}
\title{Calculate Total Pathway Probabilities of Decision Tree}
\usage{
calc.pathway_probs(osNode, FUN = "product")
}
\arguments{
\item{osNode}{object of class costeffectiveness.tree}

\item{FUN}{sum or product}
}
\value{
vector of values
}
\description{
Calculate Total Pathway Probabilities of Decision Tree
}
\examples{

}

