% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pathway_probs.R
\name{calc.riskprofile}
\alias{calc.riskprofile}
\title{Calculate Risk Profile}
\usage{
calc.riskprofile(osNode)
}
\arguments{
\item{osNode}{}
}
\description{
This function can be used to calculate the Risk Profile.
That is, the distribution of outcome values.
}
\examples{

}
\seealso{
\link{calc.pathway_probs}
}

