% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expectedValues.R
\name{calc_expectedValues.default}
\alias{calc_expectedValues.default}
\title{Calculate Expected Values for Each Node of Decision Tree}
\usage{
\method{calc_expectedValues}{default}(osNode)
}
\arguments{
\item{osNode}{}
}
\value{
osNode
}
\description{
Takes an object of class costeffectiveness_tree.
}
\seealso{
\link{costeffectiveness_tree}, \link{payoff}
}
