% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pathway_probs.R
\name{calc_pathway_probs.default}
\alias{calc_pathway_probs.default}
\title{Calculate Total Pathway Probabilities of Decision Tree}
\usage{
\method{calc_pathway_probs}{default}(osNode, ...)
}
\arguments{
\item{osNode}{object of class costeffectiveness_tree}

\item{FUN}{sum or product}
}
\value{
vector of values
}
\description{
Sequential event operations.
The probabilities are calculate with \code{FUN="product"}
and the values are calculated with \code{FUN="sum"}.
}
\seealso{
\link{calc_riskprofile}
}

