% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pathway_probs.R
\name{calc_riskprofile.costeffectiveness_tree}
\alias{calc_riskprofile.costeffectiveness_tree}
\title{Calculate Risk Profile}
\usage{
\method{calc_riskprofile}{costeffectiveness_tree}(osNode)
}
\arguments{
\item{osNode}{}
}
\description{
This function can be used to calculate the Risk Profile.
That is, the distribution of outcome values.
}
\examples{

}
\seealso{
\link{calc_pathway_probs}
}

