% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pathway_probs.R
\name{calc_riskprofile.default}
\alias{calc_riskprofile.default}
\title{Calculate Risk Profile}
\usage{
\method{calc_riskprofile}{default}(osNode, ...)
}
\arguments{
\item{osNode}{}
}
\description{
This function can be used to calculate the Risk Profile.
That is, the distribution of outcome values.
}
\examples{

}
\seealso{
\link{calc_pathway_probs}
}

