% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-effectiveness-fns2.R
\name{calcCostHealthGrid}
\alias{calcCostHealthGrid}
\title{calcCostHealthGrid}
\usage{
calcCostHealthGrid(treeCost, treeHealth, pwaytimetodiagTB = NA,
  pwaycostTB = NA, pwaytimetodiagNonTB = NA, pwaycostNonTB = NA,
  FNtime = 42, cat4percent = 75, prop_highrisk = 1, Ctest.seq = 100,
  sens.seq = 0.9, spec.seq = 0.9)
}
\arguments{
\item{spec.seq}{}
}
\description{
blah
}

