% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost-effectiveness-fns2.R
\name{CEplane_yaml}
\alias{CEplane_yaml}
\title{CEplane_yaml}
\usage{
CEplane_yaml(N = 2, tree_cost, tree_health, Ctest = 100, sens = 0.9,
  spec = 0.9, table = FALSE)
}
\arguments{
\item{table}{}
}
\description{
blah
}
\examples{

}

