% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_costeffectiveness_tree.R
\name{costeffectiveness_tree}
\alias{costeffectiveness_tree}
\title{Constructor for a Cost-Effectiveness Tree Object}
\usage{
costeffectiveness_tree(yaml_tree, details = "", data_prob = NA,
  data_val = NA, ...)
}
\arguments{
\item{yaml_tree}{YAML file or location address}

\item{details}{General details of decision tree}

\item{data_prob}{Branching probability data}

\item{data_val}{Cost or health data}

\item{...}{Additional arguments to be passed}
}
\value{
list of class costeffectiveness_object
}
\description{
The resulting object is used in the main package functions.
Branch distributions are currently permitted as uniform, gamma and triangle.
All branches require some distribution.
All branches must end in a terminal node.
}
\examples{

CEtree <- costeffectiveness_tree(yaml_tree = "raw data/LTBI_dtree-cost-distns.yaml")
print(CEtree)

CEtree$osNode <- calc.expectedValues(CEtree$osNode)
print(CEtree)

}
\seealso{
\code{\link{calc.expectedValues}}
}
