% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_costeffectiveness_tree.R
\name{create.costeffectiveness.tree}
\alias{create.costeffectiveness.tree}
\title{Create a Cost-Effectiveness Tree Object}
\usage{
create.costeffectiveness.tree(yaml_tree)
}
\arguments{
\item{yaml_tree}{YAML file or location address}
}
\value{
data.tree object of class costeffectiveness.tree
}
\description{
The resulting object is used in the main package functions.
Branch distributions are currently permitted as uniform, gamma and triangle.
All branches require some distribution.
All branches must end in a terminal node.
}
\examples{
osNode <- create.costeffectiveness.tree(yaml_tree = "raw data/LTBI_dtree-cost-distns.yaml")
print(osNode, "type", "p", "distn", "mean", "sd")

osNode <- calc.expectedValues(osNode)
print(osNode, "type", "p", "distn", "mean", "sd", "payoff")

}
\seealso{
calc.expectedValues
}

