% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IDEA-cost-effectiveness-fns-TODO.R
\name{createParamGrid}
\alias{createParamGrid}
\title{createParamGrid}
\usage{
createParamGrid(Nrepl = 10, Ctest.vec = 100, sens.vec = 0.9,
  sens.mean = NA, sens.sd = NA, spec.vec = 0.9, spec.mean = NA,
  spec.sd = NA, cat4percent.vec = 75, cat4percent.mean = NA,
  cat4percent.sd = NA, prop_highrisk.vec = 1, prop_highrisk.mean = NA,
  prop_highrisk.sd = NA, FNtime.vec = 42, FNtime.mean = NA,
  FNtime.sd = NA)
}
\arguments{
\item{FNtime.sd}{}
}
\description{
blah
}

