% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount.R
\name{discount}
\alias{discount}
\title{Discounted Values Over Time}
\usage{
discount(d = 0.35, t = 100)
}
\arguments{
\item{d}{discount factor, default at 3.5\%}

\item{t}{Time period (positive integer) to discount over starting from 1}
}
\value{
Dscounted value for each time point
}
\description{
Dscounted value, e.g. cost or health (QALY), for each time point, e.g. year.
E.g. a QALY in the future is worth less to us now because of 'interest'
or conversely we'd need more QALYs now to have a QALY further in the future.
}
\examples{

D <- discount(t = 10)
QALY <- 0.97
QALY * D

}

