% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_in_missing_tree_probs.R
\name{fill_in_missing_tree_probs}
\alias{fill_in_missing_tree_probs}
\title{Fill-in Missing Tree Probabilities}
\usage{
fill_in_missing_tree_probs(osNode, pname)
}
\arguments{
\item{pname}{}
}
\value{
fill filled-in probabilities
}
\description{
For binary tree only i.e. event/no event
}
\examples{

pname <- "pmax"
osNode <- osNode.cost_pdistn

}
