% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_state_proportions.R
\name{get_start_state_proportions}
\alias{get_start_state_proportions}
\title{Get Start State Proportions}
\usage{
get_start_state_proportions(path_probs, startstateid, samplesize, numsamples)
}
\arguments{
\item{numsamples}{}
}
\description{
Get Start State Proportions
}

