% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means_distributions.R
\name{means_distributions}
\alias{means_distributions}
\title{Means values from distributions}
\usage{
means_distributions(param.distns)
}
\arguments{
\item{param.distns}{list of distribution name and hyper-parameters}
}
\value{
double
}
\description{
Means values from distributions
}
\examples{

param.distns <- list(distn = "unif", params = c(min = 0, max = 1))
means_distributions(param.distns)

}
