% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expectedValues.R
\name{MonteCarlo.expectedValues}
\alias{MonteCarlo.expectedValues}
\title{Monte Carlo Forward Simulation of Decision Tree}
\usage{
MonteCarlo.expectedValues(osNode, n = 100)
}
\arguments{
\item{osNode}{A data.tree object with class costeffectiveness.tree}

\item{n}{Number of simulations}
}
\value{
column-named array of n sets of expected values
}
\description{
Results are returned for the nodes labelled logical in decision tree.
Require at least one logical node.
}
\examples{
## read-in decision tree
osNode <- create.costeffectiveness.tree(yaml_tree = "raw data/LTBI_dtree-cost-distns.yaml")
print(osNode, "type", "p", "distn", "mean", "sd")

## calculate a single realisation expected values
osNode <- calc.expectedValues(osNode)
print(osNode, "type", "p", "distn", "mean", "sd", "payoff")

## calculate multiple realisation for specific nodes
MonteCarlo.expectedValues(osNode, n=100)

}
\seealso{
calc.expectedValues
}

