% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_expectedValues.R
\name{MonteCarlo_expectedValues}
\alias{MonteCarlo_expectedValues}
\title{Monte Carlo Forward Simulation of Decision Tree}
\usage{
MonteCarlo_expectedValues(osNode, n)
}
\arguments{
\item{osNode}{A data.tree object with class costeffectiveness_tree}

\item{n}{Number of simulations}
}
\value{
list containing array of n sets of expected values and sampled nodes full names
}
\description{
Results are returned for the nodes labelled logical in decision tree.
Require at least one logical node.
}
\seealso{
calc_expectedValues
}

