% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarlo_ntb_screen.R
\name{MonteCarlo_n.tb_screen}
\alias{MonteCarlo_n.tb_screen}
\title{Monte Carlo samples of number of TB cases after screening}
\usage{
MonteCarlo_n.tb_screen(p_complete_Tx, n.uk_tb, n.all_tb, n = 2)
}
\arguments{
\item{p_complete_Tx}{Probability complete treatment}

\item{n.uk_tb}{Number TB cases in UK status-quo}

\item{n.all_tb}{Number of TB case in total status-quo}

\item{n}{Total sample size}
}
\value{
list
}
\description{
frequency of active tb and disease-free (screened)
TODO: generalise or remove from treeSimR::
}
\examples{
sum(IMPUTED_sample_year_cohort$all_tb)
395
sum(IMPUTED_sample_year_cohort$uk_tb)
197
}
