% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myPruneFun.R
\name{myPruneFun}
\alias{myPruneFun}
\title{Prune Tree}
\usage{
myPruneFun(x, leafName)
}
\arguments{
\item{x}{Tree node}

\item{leafName}{Character string}
}
\description{
Return just the paths with the given leaf name
}
