% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoff.R
\name{payoff}
\alias{payoff}
\title{Calculate weighted expectations}
\usage{
payoff(node)
}
\arguments{
\item{node}{}
}
\value{
jl$Do(payoff, traversal = "post-order", filterFun = isNotLeaf)
}
\description{
Calculate weighted expectations
}

