% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_costeffectiveness_tree.R
\name{print.costeffectiveness_tree}
\alias{print.costeffectiveness_tree}
\title{Print Method for Cost-Effectiveness Trees}
\usage{
\method{print}{costeffectiveness_tree}(decisiontree, ...)
}
\arguments{
\item{...}{}
}
\description{
Print Method for Cost-Effectiveness Trees
}

