% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_distributions.R
\name{sample_distributions}
\alias{sample_distributions}
\title{Sample from Standard Distributions}
\usage{
sample_distributions(param.distns)
}
\arguments{
\item{param.distns}{List of distribution names and their respective parameter values}
}
\value{
vector of sample points
}
\description{
Supply a list of defined distributions and one sample realisation is taken of each.
}

