% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoff.R, R/sample_distributions.R
\name{sampleNode}
\alias{sampleNode}
\title{Randomly Sample a Value for a Node}
\usage{
sampleNode(node)

sampleNode(node)
}
\arguments{
\item{node}{}

\item{node}{data.tree node}
}
\value{
sampled value
}
\description{
Randomly Sample a Value for a Node

Sample a data.tree Node
}

