% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_branching_values.R
\name{assign_branch_values}
\alias{assign_branch_values}
\title{Assign Branching Values to Decision Tree}
\usage{
assign_branch_values(osNode.cost, osNode.health, parameter_p = NA,
  parameter_cost = NA)
}
\arguments{
\item{osNode.cost}{data.tree object}

\item{osNode.health}{data.tree object}

\item{parameter_p}{Decision tree node probabilities}

\item{parameter_cost}{Decision tree node cost}
}
\description{
Used in deterministic sensitivity analysis.
##TODO## at present no need for parameter_health
}

