% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_riskprofile.R
\name{calc_riskprofile}
\alias{calc_riskprofile}
\alias{calc_riskprofile.default}
\alias{calc_riskprofile.costeffectiveness_tree}
\title{Calculate Risk Profile}
\usage{
calc_riskprofile(osNode)

\method{calc_riskprofile}{default}(osNode, ...)

\method{calc_riskprofile}{costeffectiveness_tree}(osNode)
}
\arguments{
\item{osNode}{}
}
\description{
Calculate the distribution of outcome values.
TODO: overlap with subset_pop_dectree()
}
\examples{

}
\seealso{
\link{calc_pathway_probs}
}
