% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_costeffectiveness_tree.R
\name{costeffectiveness_tree}
\alias{costeffectiveness_tree}
\title{Constructor for a Cost-Effectiveness Tree Object}
\usage{
costeffectiveness_tree(yaml_tree, details = "")
}
\arguments{
\item{yaml_tree}{YAML file or location address}

\item{details}{general details of decision tree}
}
\value{
data.tree object of class costeffectiveness_tree
}
\description{
The resulting object is used in the main package functions.
Branch distributions are currently permitted as uniform, gamma and triangle.
All branches require some distribution.
All branches must end in a terminal node.
}
\examples{
osNode <- costeffectiveness_tree(yaml_tree = "raw data/LTBI_dtree-cost-distns.yaml")
print(osNode, "type", "p", "distn", "mean", "sd")

osNode <- calc.expectedValues(osNode)
print(osNode, "type", "p", "distn", "mean", "sd", "payoff")

}
\seealso{
calc.expectedValues
}

