% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_start_state_proportions.R
\name{get_start_state_proportions}
\alias{get_start_state_proportions}
\title{Get Start State Proportions}
\usage{
get_start_state_proportions(path_probs, startstateid, samplesize, numsamples)
}
\arguments{
\item{path_probs}{Probabilitites along each branch, from root to leaf usinghj \link{calc_pathway_probs}}

\item{startstateid}{How to group terminal nodes}

\item{samplesize}{Total population size}

\item{numsamples}{Number of realisations}
}
\value{
list
}
\description{
Population sizes of individuals in each starting state as input for the competing risks model.
}
\examples{

}

