% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/means_distributions.R
\name{means_distributions}
\alias{means_distributions}
\title{means_distributions}
\usage{
means_distributions(param.distns)
}
\arguments{
\item{param.distns}{}
}
\description{
means_distributions
}
\examples{
means_distributions(unit_cost$aTB_TxDx)

}
