% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarlo_ntb_screen.R
\name{MonteCarlo_n.tb_screen}
\alias{MonteCarlo_n.tb_screen}
\title{MonteCarlo_n.tb_screen}
\usage{
MonteCarlo_n.tb_screen(p_complete_Tx, n.uk_tb, n.all_tb, n = 2)
}
\arguments{
\item{n}{}
}
\description{
frequency of active tb and disease-free (screened)
TODO: generalise or remove from treeSimR::
}
\examples{
sum(IMPUTED_sample_year_cohort$all_tb)
395
sum(IMPUTED_sample_year_cohort$uk_tb)
197
}
